Breaking Account.kt up into multiple files
