+++ src/commonMain/kotlin/exchange.dydx.abacus/processor/wallet/account/FillProcessor.kt
@@ -1,8 +1,8 @@
 package exchange.dydx.abacus.processor.wallet.account
 
-import exchange.dydx.abacus.output.FillLiquidity
-import exchange.dydx.abacus.output.SubaccountFill
-import exchange.dydx.abacus.output.SubaccountFillResources
+import exchange.dydx.abacus.output.account.FillLiquidity
+import exchange.dydx.abacus.output.account.SubaccountFill
+import exchange.dydx.abacus.output.account.SubaccountFillResources
 import exchange.dydx.abacus.output.input.MarginMode
 import exchange.dydx.abacus.output.input.OrderSide
 import exchange.dydx.abacus.output.input.OrderType
