+++ src/commonMain/kotlin/exchange.dydx.abacus/processor/wallet/account/HistoricalPNLsProcessor.kt
@@ -1,6 +1,6 @@
 package exchange.dydx.abacus.processor.wallet.account
 
-import exchange.dydx.abacus.output.SubaccountHistoricalPNL
+import exchange.dydx.abacus.output.account.SubaccountHistoricalPNL
 import exchange.dydx.abacus.processor.base.BaseProcessor
 import exchange.dydx.abacus.protocols.ParserProtocol
 import indexer.codegen.IndexerPnlTicksResponseObject
