+++ src/commonMain/kotlin/exchange.dydx.abacus/state/manager/notification/providers/BlockRewardNotificationProvider.kt
@@ -1,9 +1,9 @@
 package exchange.dydx.abacus.state.manager.notification.providers
 
-import exchange.dydx.abacus.output.BlockReward
 import exchange.dydx.abacus.output.Notification
 import exchange.dydx.abacus.output.NotificationPriority
 import exchange.dydx.abacus.output.NotificationType
+import exchange.dydx.abacus.output.account.BlockReward
 import exchange.dydx.abacus.state.manager.V4Environment
 import exchange.dydx.abacus.state.manager.notification.NotificationsProviderProtocol
 import exchange.dydx.abacus.state.model.TradingStateMachine
