+++ src/commonMain/kotlin/exchange.dydx.abacus/state/manager/notification/providers/OrderStatusChangesNotificationProvider.kt
@@ -3,7 +3,7 @@ package exchange.dydx.abacus.state.manager.notification.providers
 import exchange.dydx.abacus.output.Notification
 import exchange.dydx.abacus.output.NotificationPriority
 import exchange.dydx.abacus.output.NotificationType
-import exchange.dydx.abacus.output.SubaccountOrder
+import exchange.dydx.abacus.output.account.SubaccountOrder
 import exchange.dydx.abacus.output.input.OrderStatus
 import exchange.dydx.abacus.output.input.OrderType
 import exchange.dydx.abacus.protocols.ParserProtocol
