+++ src/commonMain/kotlin/exchange.dydx.abacus/state/model/TradingStateMachine+Account.kt
@@ -1,6 +1,6 @@
 package exchange.dydx.abacus.state.model
 
-import exchange.dydx.abacus.output.SubaccountOrder
+import exchange.dydx.abacus.output.account.SubaccountOrder
 import exchange.dydx.abacus.output.input.OrderStatus
 import exchange.dydx.abacus.responses.StateResponse
 import exchange.dydx.abacus.state.changes.Changes
