+++ src/commonMain/kotlin/exchange.dydx.abacus/utils/AnalyticsUtils.kt
@@ -1,6 +1,6 @@
 package exchange.dydx.abacus.utils
 
-import exchange.dydx.abacus.output.SubaccountOrder
+import exchange.dydx.abacus.output.account.SubaccountOrder
 import exchange.dydx.abacus.output.input.OrderType
 import exchange.dydx.abacus.state.manager.HumanReadableCancelOrderPayload
 import exchange.dydx.abacus.state.manager.HumanReadablePlaceOrderPayload
