+++ src/commonTest/kotlin/exchange.dydx.abacus/processor/wallet/account/OrdersProcessorTests.kt
@@ -1,6 +1,6 @@
 package exchange.dydx.abacus.processor.wallet.account
 
-import exchange.dydx.abacus.output.SubaccountOrder
+import exchange.dydx.abacus.output.account.SubaccountOrder
 import exchange.dydx.abacus.tests.mock.LocalizerProtocolMock
 import exchange.dydx.abacus.tests.mock.processor.wallet.account.OrderProcessorMock
 import exchange.dydx.abacus.utils.Parser
