+++ src/commonTest/kotlin/exchange.dydx.abacus/tests/mock/processor/wallet/account/HistoricalPNLProcessorMock.kt
@@ -1,6 +1,6 @@
 package exchange.dydx.abacus.tests.mock.processor.wallet.account
 
-import exchange.dydx.abacus.output.SubaccountHistoricalPNL
+import exchange.dydx.abacus.output.account.SubaccountHistoricalPNL
 import exchange.dydx.abacus.processor.wallet.account.HistoricalPNLProcessorProtocol
 import indexer.codegen.IndexerPnlTicksResponseObject
